requirements.txt
pystray>=0.19.0
Pillow>=9.0.0
pynput>=1.7.6
keyboard>=0.13.5
# tkinter is included in the Python standard library (usually).
# If missing, install via your OS package manager:
#   Ubuntu/Debian: sudo apt install python3-tk
#   Windows: reinstall Python and check "tcl/tk and IDLE"
